api/common.go: code shared by the server and client implementation of the Docker remote api.
Docker-DCO-1.1-Signed-off-by: Solomon Hykes <solomon@docker.com> (github: shykes)
